move gtm setting stuff into admin namespace
